Hack: Utilize no-op behavior of nvim-treesitter.install function
For some bizarre reason, the function doesn't error out when I provide a
file type that doesn't have an associated parser. It simply does
nothing. I have no idea if this is intended or not, but I'm going to use
this to the fullest to get rid of the stupid monkey patch for zsh, and
fixing the sh file type not having the parser start